requirements.txt
asttokens==3.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cmake==3.29.0.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
cytoolz==1.0.1
decorator==5.1.1
dlib @ file:///C:/Users/Adam/OneDrive/Documents/Downloads/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dustgoggles==0.7.5
empyrical==0.5.5
executing==2.1.0
Flask==3.1.0
Flask-Login==0.6.3
fonttools==4.53.1
frozendict==2.4.6
greenlet==3.1.1
html5lib==1.1
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
more-itertools==10.5.0
multidict==6.1.0
multitasking==0.0.11
numpy==2.0.1
packaging==24.1
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
pdr==1.3.0
peewee==3.17.8
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pyfolio==0.9.2
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rms-vax==1.0.5
scikit-learn==1.6.0
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
threadpoolctl==3.5.0
toolz==1.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yfinance==0.2.51
